Refactor task management service and update proto definitions
- Changed the directory structure for protobuf files, moving them to a centralized 'proto' directory.
- Updated import paths in the task service and handler to reference the new protobuf package location.
- Refactored task-related proto definitions to improve clarity and consistency, including renaming fields and updating service methods.
- Removed outdated generated files and ensured new gRPC generated files are correctly integrated.
- Enhanced unit tests to align with the updated proto definitions and service methods, ensuring robust functionality and error handling.
