+++ internal/task/service/task_service_test.go
@@ -5,9 +5,9 @@ import (
 	"testing"
 	"time"
 
+	"github.com/my-backend-project/internal/pb"
 	"github.com/my-backend-project/internal/pkg/errors"
 	"github.com/my-backend-project/internal/task/model"
-	"github.com/my-backend-project/internal/task/pb"
 
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/mock"
