--- proto/task/task.proto
@@ -1,84 +1,0 @@
-syntax = "proto3";
-
-package task;
-
-option go_package = "bnp/internal/task/pb";
-
-import "google/protobuf/timestamp.proto";
-import "google/protobuf/empty.proto";
-
-enum TaskStatus {
-  TASK_STATUS_UNSPECIFIED = 0;
-  TASK_STATUS_PENDING = 1;
-  TASK_STATUS_ACTIVE = 2;
-  TASK_STATUS_COMPLETE = 3;
-}
-
-message Task {
-  string task_id = 1;
-  string user_id = 2;
-  string title = 3;
-  string description = 4;
-  TaskStatus status = 5;
-  google.protobuf.Timestamp due_date = 6;
-  google.protobuf.Timestamp created_at = 7;
-  google.protobuf.Timestamp updated_at = 8;
-}
-
-message CreateTaskRequest {
-  string user_id = 1;
-  string title = 2;
-  string description = 3;
-  TaskStatus status = 4;
-  google.protobuf.Timestamp due_date = 5;
-}
-
-message CreateTaskResponse {
-  string task_id = 1;
-}
-
-message GetTaskRequest {
-  string task_id = 1;
-}
-
-message GetTaskResponse {
-  Task task = 1;
-}
-
-message ListTasksRequest {
-  string user_id = 1;
-  TaskStatus status = 2;
-  int32 page_size = 3;
-  string page_token = 4;
-}
-
-message ListTasksResponse {
-  repeated Task tasks = 1;
-  string next_page_token = 2;
-  int32 total_count = 3;
-}
-
-message UpdateTaskRequest {
-  string task_id = 1;
-  string user_id = 2;
-  string title = 3;
-  string description = 4;
-  TaskStatus status = 5;
-  google.protobuf.Timestamp due_date = 6;
-}
-
-message UpdateTaskResponse {
-  Task task = 1;
-}
-
-message DeleteTaskRequest {
-  string task_id = 1;
-}
-
-service TaskService {
-  rpc CreateTask(CreateTaskRequest) returns (CreateTaskResponse);
-  rpc GetTask(GetTaskRequest) returns (GetTaskResponse);
-  rpc ListTasks(ListTasksRequest) returns (ListTasksResponse);
-  rpc UpdateTask(UpdateTaskRequest) returns (UpdateTaskResponse);
-  rpc DeleteTask(DeleteTaskRequest) returns (google.protobuf.Empty);
-}
